login: move to model directory
