+++ app/generate-password.go
@@ -4,13 +4,13 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/pass-wall/passwall-server/api/login"
 	"github.com/pass-wall/passwall-server/internal/encryption"
+	"github.com/pass-wall/passwall-server/model"
 )
 
 // GeneratePassword generates new password
 func GeneratePassword(c *gin.Context) {
 	password := encryption.Password()
-	response := login.LoginResponse{"Success", password}
+	response := model.LoginResponse{"Success", password}
 	c.JSON(http.StatusOK, response)
 }
