+++ model/login_helper.go
@@ -1,4 +1,4 @@
-package login
+package model
 
 import (
 	"encoding/base64"
