+++ model/login_mapper.go
@@ -1,4 +1,4 @@
-package login
+package model
 
 import (
 	"strings"
